% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_database.R
\name{create.plot.images}
\alias{create.plot.images}
\title{Create daily static plot images}
\usage{
create.plot.images(
  db.file = "netscan-test.db",
  plot.width = 1000,
  plot.height = 1000
)
}
\arguments{
\item{db.file}{Character. File path and name of the SQLite database to update.}

\item{plot.width}{Numeric. Width of the plot image, in pixels.}

\item{plot.height}{Numeric. height of the plot image, in pixels.}
}
\value{
NULL (invisible)
}
\description{
Create daily static plot images
}
\examples{
1
}
