% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_update_database.R
\name{database.update}
\alias{database.update}
\title{Update netscan database}
\usage{
database.update(
  db.file = "netscan-test.db",
  confirm.rpc = TRUE,
  query.asn = TRUE,
  get.domains = TRUE
)
}
\arguments{
\item{db.file}{Character. File path and name of the SQLite database to update.}

\item{confirm.rpc}{Logical. Whether to test the RPC ports of the nodes
that claim to have them available.}

\item{query.asn}{Logical. Whether to query Team Cymru for the Autonomous
System Number (ASN) of each node.}

\item{get.domains}{Logical. Whether to query
\url{https://xmr.ditatompel.com/remote-nodes} for the domain names
associated with the remote nodes.}
}
\value{
NULL (invisible)
}
\description{
Update netscan database
}
\examples{
1
}
