% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmrnetscan-package.R
\docType{package}
\name{xmrnetscan-package}
\alias{xmrnetscan}
\alias{xmrnetscan-package}
\title{xmrnetscan: Shiny App with with Monero Node Data Collected from Network Scans}
\description{
Shiny App with Monero Node Data Collected from Network Scans.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Rucknium/xmrnetscan}
  \item Report bugs at \url{https://github.com/Rucknium/xmrnetscan/issues}
}

}
\author{
\strong{Maintainer}:  Rucknium \email{Rucknium@protonmail.com} (\href{https://orcid.org/0000-0001-5999-8950}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item  Boog900 (Author of Rust network crawler) [copyright holder]
}

}
\keyword{internal}
